requirements.txt
python-telegram-bot
requests
Pillow

